DEMRUM-5290: Split NavigationEventProcessorTests to fix SwiftLint violations
* Extract test processors to MockNavigationEventProcessors.swift
* Move edge-case tests to NavigationEventProcessorEdgeCaseTests.swift
* Resolves file_length, type_body_length, and function_body_length
  violations
