+++ SplunkNavigation/Tests/SplunkNavigationTests/NavigationEventProcessorTests.swift
@@ -233,137 +233,6 @@ final class NavigationEventProcessorTests: XCTestCase {
     }
 
 
-    // MARK: - Mid-transition suppression cleanup
-
-    func testDidShowSuppressionCleansUpNavigationPair() async {
-        let fixture = makeNavigationStreamFixture(
-            navigationEventProcessor: AllowThenSuppressProcessor()
-        )
-
-        defer {
-            fixture.finish()
-        }
-
-        let navigation = fixture.navigation
-        navigation.preferences.enableAutomatedTracking = true
-        navigation.startDetection()
-
-        let controllerIdentifier = ObjectIdentifier(NSString())
-        let navigationControllerIdentifier = ObjectIdentifier(NSNumber(value: 50))
-
-        fixture.sendTransition(
-            type: .navigationControllerWillShow,
-            navigationControllerIdentifier: navigationControllerIdentifier,
-            controllerIdentifier: controllerIdentifier,
-            controllerTypeName: "TransientViewController"
-        )
-
-        // Wait for willShow to be processed and NavigationPair to be stored
-        try? await Task.sleep(nanoseconds: 200_000_000)
-
-        fixture.sendTransition(
-            type: .navigationControllerDidShow,
-            navigationControllerIdentifier: navigationControllerIdentifier,
-            controllerIdentifier: controllerIdentifier,
-            controllerTypeName: "TransientViewController"
-        )
-
-        // Wait for didShow suppression to clean up
-        try? await Task.sleep(nanoseconds: 200_000_000)
-
-        let pair = await navigation.model.navigation(for: controllerIdentifier)
-        XCTAssertNil(pair, "NavigationPair should be removed when processor suppresses at didShow")
-    }
-
-
-    // MARK: - Reserved span key protection
-
-    func testReservedSpanKeysCannotBeOverriddenByProcessor() async {
-        // Use a processor that returns attributes containing all SDK-reserved keys.
-        let fixture = makeNavigationStreamFixture(
-            navigationEventProcessor: ReservedKeyOverrideProcessor()
-        )
-
-        defer {
-            fixture.finish()
-        }
-
-        let navigation = fixture.navigation
-        navigation.preferences.enableAutomatedTracking = true
-        navigation.startDetection()
-
-        // Set an initial screen via manual tracking so last.screen.name has a known value.
-        navigation.track(screen: "HomeScreen")
-
-        let didSetInitial = await waitUntil {
-            await navigation.model.screenName == "HomeScreen"
-        }
-        XCTAssertTrue(didSetInitial)
-
-        // Trigger an automated navigation event through the processor.
-        let controllerIdentifier = ObjectIdentifier(NSString())
-        let navigationControllerIdentifier = ObjectIdentifier(NSNumber(value: 99))
-
-        fixture.sendTransition(
-            type: .navigationControllerWillShow,
-            navigationControllerIdentifier: navigationControllerIdentifier,
-            controllerIdentifier: controllerIdentifier,
-            controllerTypeName: "SettingsViewController"
-        )
-        fixture.sendTransition(
-            type: .navigationControllerDidShow,
-            navigationControllerIdentifier: navigationControllerIdentifier,
-            controllerIdentifier: controllerIdentifier,
-            controllerTypeName: "SettingsViewController"
-        )
-
-        // The processor returns name: "SettingsViewController" with attributes
-        // attempting to set screen.name, last.screen.name, component, and
-        // navigation.name to "hacker". Verify the SDK uses the processor's
-        // `name` field for the screen name, not the "screen.name" attribute.
-        let didUpdate = await waitUntil {
-            await navigation.model.screenName == "SettingsViewController"
-        }
-        XCTAssertTrue(didUpdate)
-
-        let currentScreenName = await navigation.model.screenName
-        XCTAssertEqual(
-            currentScreenName,
-            "SettingsViewController",
-            "Screen name must come from NavigationEvent.name, not from the 'screen.name' attribute"
-        )
-        XCTAssertNotEqual(
-            currentScreenName,
-            "hacker",
-            "Processor attributes must not override the SDK-computed screen name"
-        )
-
-        // Verify the processor returns an event that carries the override attributes.
-        // Navigation+Span.swift writes these user attributes BEFORE the SDK-reserved
-        // keys (component, navigation.name, screen.name, last.screen.name) so that
-        // clearAndSetAttribute for SDK keys always wins.
-        let processedEvent = await navigation.processAutomatedNavigationEvent(
-            "SettingsViewController",
-            controllerIdentifier: controllerIdentifier
-        )
-        XCTAssertNotNil(processedEvent)
-        XCTAssertEqual(processedEvent?.name, "SettingsViewController")
-        XCTAssertNotNil(
-            processedEvent?.attributes,
-            "Processor should return attributes (including reserved key overrides)"
-        )
-
-        // Confirm the processor did return attributes that attempt to override reserved keys.
-        // The span-level protection is structural: send(screenName:lastScreenName:start:attributes:)
-        // in Navigation+Span.swift writes user attributes first, then overwrites with SDK values.
-        let attrs = processedEvent?.attributes
-        XCTAssertEqual(attrs?["screen.name"] as? String, "hacker")
-        XCTAssertEqual(attrs?["component"] as? String, "hacker")
-        XCTAssertEqual(attrs?["navigation.name"] as? String, "hacker")
-        XCTAssertEqual(attrs?["last.screen.name"] as? String, "hacker")
-    }
-
-
     // MARK: - Event suppression
 
     func testProcessorNilSuppressesNavigation() async {
@@ -411,65 +280,5 @@ final class NavigationEventProcessorTests: XCTestCase {
         // Screen name should remain unchanged because the event was suppressed
         let currentScreenName = await navigation.model.screenName
         XCTAssertEqual(currentScreenName, "InitialScreen")
-    }
-}
-
-
-// MARK: - Test processors
-
-private final class PrefixingProcessor: NavigationEventProcessor {
-    let prefix: String
-
-    init(prefix: String) {
-        self.prefix = prefix
-    }
-
-    func onViewController(typeName: String, controllerIdentity _: String) -> NavigationEvent? {
-        NavigationEvent(name: "\(prefix)/\(typeName)")
-    }
-}
-
-private final class RejectingProcessor: NavigationEventProcessor {
-    func onViewController(typeName _: String, controllerIdentity _: String) -> NavigationEvent? {
-        NavigationEvent(name: "REJECTED")
-    }
-}
-
-private final class SuppressingProcessor: NavigationEventProcessor {
-    func onViewController(typeName _: String, controllerIdentity _: String) -> NavigationEvent? {
-        nil
-    }
-}
-
-private final class AllowThenSuppressProcessor: NavigationEventProcessor {
-    private let lock = NSLock()
-    private var callCount = 0
-
-    func onViewController(typeName: String, controllerIdentity _: String) -> NavigationEvent? {
-        lock.lock()
-        defer { lock.unlock() }
-
-        callCount += 1
-
-        if callCount == 1 {
-            return NavigationEvent(name: typeName)
-        }
-
-        return nil
-    }
-}
-
-/// Returns the event with attributes that attempt to override all SDK-reserved span keys.
-private final class ReservedKeyOverrideProcessor: NavigationEventProcessor {
-    func onViewController(typeName: String, controllerIdentity _: String) -> NavigationEvent? {
-        NavigationEvent(
-            name: typeName,
-            attributes: [
-                "screen.name": "hacker",
-                "last.screen.name": "hacker",
-                "component": "hacker",
-                "navigation.name": "hacker"
-            ]
-        )
     }
 }
